requirements.txt
openai
python-dotenv
streamlit
py -3.11